update：容量规划

+++ internal/entity/server_cap_planning.go
@@ -5,7 +5,7 @@ const ServerCapPlanningTable = "server_cap_planning"
 type ServerCapPlanning struct {
 	PlanId             int64 `gorm:"column:plan_id" json:"planId"`                          // 方案id
 	CapacityBaselineId int64 `gorm:"column:capacity_baseline_id" json:"capacityBaselineId"` // 容量指标id
-	Number             int64 `gorm:"column:number" json:"number"`                           // 数量
+	Number             int   `gorm:"column:number" json:"number"`                           // 数量
 }
 
 func (entity *ServerCapPlanning) TableName() string {
